Remove nn_utils.py file; move pad_zeros() function to common_utils.py
