+++ mimic3models/decompensation/utils.py
@@ -1,7 +1,6 @@
 from __future__ import absolute_import
 from __future__ import print_function
 
-from mimic3models import nn_utils
 from mimic3models import common_utils
 import threading
 import os
@@ -58,7 +57,7 @@ class BatchGen(object):
                 (Xs, ys, ts, names) = common_utils.sort_and_shuffle([Xs, ys, ts, names], B)
 
                 for i in range(0, current_size, B):
-                    X = nn_utils.pad_zeros(Xs[i:i + B])
+                    X = common_utils.pad_zeros(Xs[i:i + B])
                     y = np.array(ys[i:i + B])
                     batch_names = names[i:i+B]
                     batch_ts = ts[i:i+B]
@@ -176,9 +175,9 @@ class BatchGenDeepSupervision(object):
                 names = self.names[i:i + B]
                 ts = self.ts[i:i + B]
 
-                X = nn_utils.pad_zeros(X)  # (B, T, D)
-                mask = nn_utils.pad_zeros(mask)  # (B, T)
-                y = nn_utils.pad_zeros(y)
+                X = common_utils.pad_zeros(X)  # (B, T, D)
+                mask = common_utils.pad_zeros(mask)  # (B, T)
+                y = common_utils.pad_zeros(y)
                 y = np.expand_dims(y, axis=-1)  # (B, T, 1)
                 batch_data = ([X, mask], y)
                 if not self.return_names:
