+++ mimic3models/length_of_stay/utils.py
@@ -3,7 +3,6 @@ from __future__ import print_function
 
 from mimic3models import metrics
 from mimic3models import common_utils
-from mimic3models import nn_utils
 import threading
 import os
 import numpy as np
@@ -60,7 +59,7 @@ class BatchGen(object):
                 (Xs, ys, ts, names) = common_utils.sort_and_shuffle([Xs, ys, ts, names], B)
 
                 for i in range(0, current_size, B):
-                    X = nn_utils.pad_zeros(Xs[i:i+B])
+                    X = common_utils.pad_zeros(Xs[i:i + B])
                     y = ys[i:i+B]
                     y_true = np.array(y)
                     batch_names = names[i:i+B]
@@ -194,7 +193,7 @@ class BatchGenDeepSupervision(object):
                 ts = self.ts[i:i+B]
 
                 y_true = [np.array(x) for x in y]
-                y_true = nn_utils.pad_zeros(y_true)
+                y_true = common_utils.pad_zeros(y_true)
                 y_true = np.expand_dims(y_true, axis=-1)
 
                 if self.partition == 'log':
@@ -202,9 +201,9 @@ class BatchGenDeepSupervision(object):
                 if self.partition == 'custom':
                     y = [np.array([metrics.get_bin_custom(x, 10) for x in z]) for z in y]
 
-                X = nn_utils.pad_zeros(X)  # (B, T, D)
-                mask = nn_utils.pad_zeros(mask)  # (B, T)
-                y = nn_utils.pad_zeros(y)
+                X = common_utils.pad_zeros(X)  # (B, T, D)
+                mask = common_utils.pad_zeros(mask)  # (B, T)
+                y = common_utils.pad_zeros(y)
                 y = np.expand_dims(y, axis=-1)
 
                 if self.return_y_true:
