+++ mimic3models/multitask/utils.py
@@ -2,7 +2,6 @@ from __future__ import absolute_import
 from __future__ import print_function
 
 from mimic3models import metrics
-from mimic3models import nn_utils
 from mimic3models import common_utils
 import numpy as np
 import threading
@@ -138,7 +137,7 @@ class BatchGen(object):
 
                 # X
                 X = self.data['X'][i:i+B]
-                X = nn_utils.pad_zeros(X, min_length=self.ihm_pos+1)
+                X = common_utils.pad_zeros(X, min_length=self.ihm_pos + 1)
                 T = X.shape[1]
 
                 # ihm
@@ -153,23 +152,23 @@ class BatchGen(object):
 
                 # decomp
                 decomp_M = self.data['decomp_M'][i:i+B]
-                decomp_M = nn_utils.pad_zeros(decomp_M, min_length=self.ihm_pos+1)
+                decomp_M = common_utils.pad_zeros(decomp_M, min_length=self.ihm_pos + 1)
                 decomp_y = self.data['decomp_y'][i:i+B]
-                decomp_y = nn_utils.pad_zeros(decomp_y, min_length=self.ihm_pos+1)
+                decomp_y = common_utils.pad_zeros(decomp_y, min_length=self.ihm_pos + 1)
                 decomp_y = np.expand_dims(decomp_y, axis=-1)  # (B, T, 1)
                 outputs.append(decomp_y)
 
                 # los
                 los_M = self.data['los_M'][i:i+B]
-                los_M = nn_utils.pad_zeros(los_M, min_length=self.ihm_pos+1)
+                los_M = common_utils.pad_zeros(los_M, min_length=self.ihm_pos + 1)
                 los_y = self.data['los_y'][i:i+B]
-                los_y_true = nn_utils.pad_zeros(los_y, min_length=self.ihm_pos+1)
+                los_y_true = common_utils.pad_zeros(los_y, min_length=self.ihm_pos + 1)
 
                 if self.partition == 'log':
                     los_y = [np.array([metrics.get_bin_log(x, 10) for x in z]) for z in los_y]
                 if self.partition == 'custom':
                     los_y = [np.array([metrics.get_bin_custom(x, 10) for x in z]) for z in los_y]
-                los_y = nn_utils.pad_zeros(los_y, min_length=self.ihm_pos+1)
+                los_y = common_utils.pad_zeros(los_y, min_length=self.ihm_pos + 1)
                 los_y = np.expand_dims(los_y, axis=-1)  # (B, T, 1)
                 outputs.append(los_y)
